// src/app/models/Member.ts
export interface Member {
    id: number;
    device_id: string;
    phone_number: string;
    birth_day: string;
    birth_time: string;
    status: string;
    type: string;
    active_at: string;
    expired_on: string;
    token: string;
    is_deleted: number;
    created_at: string;
    updated_at: string;
}